+++ CHANGELOG.md
@@ -13,6 +13,16 @@
 - **Reactive options no longer support observables.** They now use **signals**, so you'll need to provide a signal or a function that returns a signal.
 - The **action callback** now receives a single argument: `ActionButtonEvent`, which contains both the original event and a reference to the component.
 
+### Table footer improvements
+
+- Footer cells can now handle signals, making async data possible without workarounds
+- Footer cells can display a loading spinner while async data is loading
+- Roadmap: would be nice to also have footer renderers.
+
+#### Breaking Changes
+- All deprecated `TableCell` properties have been removed as the logic behind them was already removed.
+
+
 ## 19.0.3
 
 - Fixed issues with table footers not showing up